Use Paint to support gradients in vector drawables
