+++ vgo-core/src/main/kotlin/com/jzbrooks/vgo/core/optimization/RemoveTransparentPaths.kt
@@ -1,5 +1,7 @@
 package com.jzbrooks.vgo.core.optimization
 
+import com.jzbrooks.vgo.core.Color
+import com.jzbrooks.vgo.core.Paint
 import com.jzbrooks.vgo.core.graphic.ClipPath
 import com.jzbrooks.vgo.core.graphic.ContainerElement
 import com.jzbrooks.vgo.core.graphic.Extra
@@ -32,7 +34,10 @@ class RemoveTransparentPaths : TopDownOptimization {
                     // Colors that aren't able to be parsed may remain in the foreign map
                     element.foreign.keys.any { it.contains("color", ignoreCase = true) } ||
                     // If a path isn't transparent, allow it
-                    (element.fill.alpha != 0.toUByte() || element.stroke.alpha != 0.toUByte())
+                    !element.fill.isTransparentColor() ||
+                    !element.stroke.isTransparentColor()
             }
     }
+
+    private fun Paint.isTransparentColor(): Boolean = this is Color && alpha == 0.toUByte()
 }
