+++ vgo-core/src/test/kotlin/com/jzbrooks/vgo/core/ColorTest.kt
@@ -8,18 +8,18 @@ class ColorTest {
     @Test
     fun testArgbHexString() {
         val color = Color(0x44ff1133u)
-        assertThat(color.toHexString(Color.HexFormat.ARGB)).isEqualTo("#44ff1133")
+        assertThat(color.toHexString(HexFormat.ARGB)).isEqualTo("#44ff1133")
     }
 
     @Test
     fun testRgbaHexString() {
         val color = Color(0x44ff1133u)
-        assertThat(color.toHexString(Color.HexFormat.RGBA)).isEqualTo("#ff113344")
+        assertThat(color.toHexString(HexFormat.RGBA)).isEqualTo("#ff113344")
     }
 
     @Test
     fun testShortenedHex() {
         val color = Color(0xffffffffu)
-        assertThat(color.toHexString(Color.HexFormat.ARGB)).isEqualTo("#fff")
+        assertThat(color.toHexString(HexFormat.ARGB)).isEqualTo("#fff")
     }
 }
