+++ vgo-core/src/test/kotlin/com/jzbrooks/vgo/core/util/element/ElementConstructors.kt
@@ -1,7 +1,7 @@
 package com.jzbrooks.vgo.core.util.element
 
-import com.jzbrooks.vgo.core.Color
 import com.jzbrooks.vgo.core.Colors
+import com.jzbrooks.vgo.core.Paint
 import com.jzbrooks.vgo.core.graphic.Element
 import com.jzbrooks.vgo.core.graphic.ElementVisitor
 import com.jzbrooks.vgo.core.graphic.Graphic
@@ -24,9 +24,9 @@ fun createPath(
     commands: List<Command> = emptyList(),
     id: String? = null,
     foreign: MutableMap<String, String> = mutableMapOf(),
-    fill: Color = Colors.BLACK,
+    fill: Paint = Colors.BLACK,
     fillRule: Path.FillRule = Path.FillRule.NON_ZERO,
-    stroke: Color = Colors.TRANSPARENT,
+    stroke: Paint = Colors.TRANSPARENT,
     strokeWidth: Float = 1f,
     strokeLineCap: Path.LineCap = Path.LineCap.BUTT,
     strokeLineJoin: Path.LineJoin = Path.LineJoin.MITER,
