+++ vgo/src/test/resources/baseline/vgo_optimized.xml
@@ -1,0 +1,32 @@
+<vector xmlns:aapt="http://schemas.android.com/aapt" xmlns:android="http://schemas.android.com/apk/res/android" android:height="514dp" android:viewportHeight="514" android:viewportWidth="800" android:width="800dp">
+  <group>
+    <clip-path android:pathData="M140.62,354.14l-70.67-210.1h50.04l40.5,127.59c2.54,7.64 4.77,15.34 6.68,23.11 1.91,7.76 3.25,14.45 4.01,20.05 1.02-5.6 2.36-12.29 4.01-20.05 1.66-7.77 3.63-15.35 5.92-22.73l40.11-127.97h50.43l-70.29,210.1h-60.74Zm202.84,68.76v-39.73h59.21c11.21,0 19.87-2.99 25.98-8.98 6.11-5.98 9.17-14.07 9.17-24.25v-12.61l0.76-32.85c-4.33,11.71-11.27,20.82-20.82,27.31-9.55,6.5-21.2,9.74-34.95,9.74-21.39,0-38.39-7.19-51-21.58-12.6-14.39-18.91-33.94-18.91-58.64v-40.87c0-24.45 6.31-43.93 18.91-58.45 12.61-14.51 29.61-21.77 51-21.77 13.75,0 25.4,3.25 34.95,9.74 9.55,6.49 16.49,15.72 20.82,27.69v-33.61h46.61v207.04c0,22.16-7.33,39.67-21.97,52.53-14.64,12.86-34.57,19.29-59.78,19.29h-59.98Zm55.39-122.24c12.23,0 21.71-3.63 28.46-10.89 6.75-7.26 10.12-17.38 10.12-30.37v-36.67c0-12.99-3.37-23.11-10.12-30.37s-16.23-10.89-28.46-10.89c-12.22,0-21.64,3.51-28.26,10.51-6.63,7-9.94,16.74-9.94,29.22v39.73c0,12.48 3.31,22.22 9.94,29.22 6.62,7.01 16.04,10.51 28.26,10.51Zm230.35,56.92c-27,0-48.39-7.58-64.18-22.73-15.79-15.16-23.68-35.59-23.68-61.31v-48.9c0-25.98 7.89-46.48 23.68-61.5 15.79-15.03 37.18-22.54 64.18-22.54 26.99,0 48.38,7.51 64.17,22.54 15.79,15.02 23.69,35.52 23.69,61.5v48.9c0,25.72-7.9,46.15-23.69,61.31-15.79,15.15-37.18,22.73-64.17,22.73Zm0-41.64c12.73,0 22.6-3.5 29.6-10.51 7.01-7 10.51-17.12 10.51-30.37v-51.95c0-13.24-3.5-23.36-10.51-30.37-7-7-16.87-10.5-29.6-10.5-12.74,0-22.6,3.5-29.61,10.5-7,7.01-10.5,17.13-10.5,30.37v51.95c0,13.25 3.5,23.37 10.5,30.37 7.01,7.01 16.87,10.51 29.61,10.51Z"/>
+    <path android:pathData="M140.62,354.14l-70.67-210.1h50.04l40.5,127.59a262.27,262.27,0,0,1,6.87,22.92c1.91,8.02 3.06,14.51 3.82,20.24 1.15-5.73 2.29-12.22 4.2-20.24 1.53-7.64 3.44-15.28 5.73-22.54l40.11-127.97h50.43l-70.29,210.1h-60.74M343.46,422.9v-39.73h59.21c11.08,0 19.87-3.06 25.98-8.79 6.11-6.11 9.17-14.13 9.17-24.44v-12.61l0.76-32.85c-4.2,11.84-11.08,20.63-20.63,27.12-9.55,6.49-21.39,9.93-35.14,9.93-21.39,0-38.2-7.26-50.81-21.77-12.6-14.14-19.1-33.62-19.1-58.45v-40.87c0-24.45 6.5-43.93 19.1-58.45 12.61-14.51 29.42-21.77 50.81-21.77 13.75,0 25.59,3.05 35.14,9.55 9.55,6.49 16.43,16.04 20.63,27.88v-33.61h46.61v207.04c0,22.16-7.26,39.73-21.78,52.72-14.9,12.6-34.76,19.1-59.97,19.1h-59.98Zm55.39-122.24c12.23,0 21.78-3.82 28.65-11.08 6.5-7.26 9.94-17.19 9.94-30.18v-36.67c0-12.99-3.44-23.3-9.94-30.56-6.87-7.26-16.42-10.7-28.65-10.7-12.22,0-21.77,3.44-28.26,10.32-6.5,7.26-9.94,16.81-9.94,29.41v39.73c0,12.61 3.44,22.16 9.94,29.03 6.49,7.26 16.04,10.7 28.26,10.7ZM629.2,357.58c-27.12,0-48.51-7.64-64.18-22.92-15.66-14.9-23.68-35.53-23.68-61.12v-48.9c0-25.98 8.02-46.6 23.68-61.5 15.67-14.9 37.06-22.54 64.18-22.54s48.51,7.64 64.18,22.54c15.66,14.9 23.68,35.52 23.68,61.5v48.9c0,25.59-8.02,46.22-23.68,61.12-15.67,15.28-37.06,22.92-64.18,22.92Zm0-41.64c12.61,0 22.54-3.44 29.79-10.7 6.88-6.87 10.32-16.81 10.32-30.18v-51.95c0-13.37-3.44-23.3-10.32-30.56-7.25-6.87-17.18-10.31-29.79-10.31s-22.54,3.44-29.41,10.31c-7.26,7.26-10.7,17.19-10.7,30.56v51.95c0,13.37 3.44,23.31 10.7,30.18 6.87,7.26 16.8,10.7 29.41,10.7Z">
+      <aapt:attr name="android:fillColor">
+        <gradient android:endX="-71.24" android:endY="35.26" android:startX="-72.24" android:startY="35.26" android:type="linear">
+          <item android:color="#b125ea" android:offset="0"/>
+          <item android:color="#833fef" android:offset="0.49"/>
+          <item android:color="#008aff" android:offset="1"/>
+        </gradient>
+      </aapt:attr>
+    </path>
+    <path android:pathData="M140.62,354.14l-70.67-210.1h50.04l40.49,127.59c2.55,7.64 4.78,15.34 6.69,23.11 1.91,7.76 3.25,14.45 4.01,20.05 1.02-5.6 2.36-12.29 4.01-20.05 1.66-7.77 3.63-15.35 5.92-22.73l40.11-127.97h50.43l-70.29,210.1h-60.74Zm202.84,68.76v-39.73h59.21c11.21,0 19.87-2.99 25.98-8.98 6.11-5.98 9.17-14.07 9.17-24.25v-12.61l0.76-32.85c-4.33,11.71-11.27,20.82-20.82,27.31-9.55,6.5-21.2,9.74-34.95,9.74-21.39,0-38.39-7.19-51-21.58-12.6-14.39-18.91-33.93-18.91-58.64v-40.87c0-24.45 6.31-43.93 18.91-58.45 12.61-14.51 29.61-21.77 51-21.77 13.75,0 25.4,3.24 34.95,9.74 9.55,6.49 16.49,15.72 20.82,27.69v-33.61h46.6v207.04c0,22.16-7.32,39.67-21.96,52.53-14.64,12.86-34.57,19.29-59.78,19.29h-59.98Zm55.39-122.24c12.23,0 21.71-3.63 28.46-10.89 6.75-7.26 10.12-17.38 10.12-30.37v-36.67c0-12.99-3.37-23.11-10.12-30.37s-16.23-10.89-28.46-10.89c-12.22,0-21.64,3.51-28.27,10.51-6.62,7-9.93,16.74-9.93,29.22v39.73c0,12.48 3.31,22.22 9.93,29.22 6.63,7.01 16.05,10.51 28.27,10.51Zm230.35,56.92c-27,0-48.39-7.58-64.18-22.73-15.79-15.16-23.68-35.59-23.68-61.31v-48.9c0-25.98 7.89-46.48 23.68-61.5 15.79-15.03 37.18-22.54 64.18-22.54 26.99,0 48.39,7.51 64.17,22.54 15.79,15.02 23.69,35.52 23.69,61.5v48.9c0,25.72-7.9,46.15-23.69,61.31-15.78,15.15-37.18,22.73-64.17,22.73Zm0-41.64c12.73,0 22.6-3.5 29.6-10.51 7.01-7 10.51-17.12 10.51-30.37v-51.95c0-13.24-3.5-23.36-10.51-30.37-7-7-16.87-10.5-29.6-10.5-12.73,0-22.6,3.5-29.61,10.5-7,7.01-10.5,17.13-10.5,30.37v51.95c0,13.25 3.5,23.37 10.5,30.37 7.01,7.01 16.88,10.51 29.61,10.51Z">
+      <aapt:attr name="android:fillColor">
+        <gradient android:endX="743.8" android:endY="75.28" android:startX="56.2" android:startY="75.28" android:type="linear">
+          <item android:color="#b125ea" android:offset="0"/>
+          <item android:color="#9441ef" android:offset="0.51"/>
+          <item android:color="#3a98ff" android:offset="1"/>
+        </gradient>
+      </aapt:attr>
+    </path>
+  </group>
+  <path android:pathData="M484.67,385.72l0.17-240.93" android:strokeColor="#0044ff" android:strokeWidth="2.59"/>
+  <path android:fillColor="#fff" android:pathData="M492.13,350.17v7.5c0,2.07-1.68,3.75-3.75,3.75h-7.5c-2.07,0-3.75-1.68-3.75-3.75v-7.5c0-2.07 1.68-3.75 3.75-3.75h7.5c2.07,0 3.75,1.68 3.75,3.75M492.57,140.07v7.5c0,2.07-1.68,3.75-3.75,3.75h-7.5c-2.07,0-3.75-1.68-3.75-3.75v-7.5c0-2.07 1.68-3.75 3.75-3.75h7.5c2.07,0 3.75,1.68 3.75,3.75" android:strokeColor="#0044ff" android:strokeWidth="3.82"/>
+  <path android:fillColor="#0044ff" android:pathData="M484.77,386.75m-4,0a4,4,135,1,1,8,-0a4,4,135,1,1,-8,0" android:strokeColor="#0044ff" android:strokeWidth="0.52"/>
+  <path android:pathData="M335.17,140.02l76.5,0.08" android:strokeColor="#0044ff" android:strokeWidth="2.56"/>
+  <path android:fillColor="#fff" android:pathData="M371.46,139.07a7.5,7.5,89.99,1,0,15,0a7.5,7.5,89.99,1,0,-15,0" android:strokeColor="#0044ff" android:strokeWidth="3.12"/>
+  <path android:fillColor="#0044ff" android:pathData="M333.72,140.07m-4,0a4,4,135,1,1,8,0a4,4,135,1,1,-8,0M411.96,140.07m-4,0a4,4,135,1,1,8,0a4,4,135,1,1,-8,0" android:strokeColor="#0044ff" android:strokeWidth="0.52"/>
+  <path android:pathData="M344.34,422.47l99.03,0.04" android:strokeColor="#0044ff" android:strokeWidth="3.62"/>
+  <path android:fillColor="#fff" android:pathData="M416.57,418.74v7.5c0,2.07-1.68,3.75-3.75,3.75h-7.5c-2.07,0-3.75-1.68-3.75-3.75v-7.5c0-2.07 1.68-3.75 3.75-3.75h7.5c2.07,0 3.75,1.68 3.75,3.75M351.23,418.74v7.5c0,2.07-1.68,3.75-3.75,3.75h-7.5c-2.07,0-3.75-1.68-3.75-3.75v-7.5c0-2.07 1.68-3.75 3.75-3.75h7.5c2.07,0 3.75,1.68 3.75,3.75" android:strokeColor="#0044ff" android:strokeWidth="3.82"/>
+  <path android:fillColor="#0044ff" android:pathData="M443.15,422.49m-4,0a4,4,0,1,1,8,0a4,4,0,1,1,-8,0" android:strokeColor="#0044ff" android:strokeWidth="0.52"/>
+</vector>
