+++ vgo/src/test/resources/gradient_linear.xml
@@ -1,0 +1,21 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path android:pathData="M0,0L24,0L24,24L0,24Z">
+    <aapt:attr name="android:fillColor">
+      <gradient
+          android:startX="0"
+          android:startY="0"
+          android:endX="24"
+          android:endY="0"
+          android:type="linear">
+        <item android:offset="0" android:color="#FFB125EA"/>
+        <item android:offset="0.5" android:color="#FF833FEF"/>
+        <item android:offset="1" android:color="#FF008AFF"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+</vector>
